requirements.txt
markdown2==2.3.4
django_pagedown==0.1.3
Django==1.10
django_markdown==0.8.4
